Fixed Transparent Image Bug.
Problem: If you are using the inner zoom mode and you
want to zoom a transparent PNG image. The image is always shown in the
background of the zoomed image.

Solution: Set the opacity of the source element
to 0 while zooming. By the mouseleave callback set the opacity back to 1.